Album add image (#20)
* Add an add button in the GUI to add new images

* Add an addImage method to add new images

* Fix status label alignment

Within an HBox the status label can not be positioned in the center
(vertically and horizontally). A VBox resolves this issue.

* Remove unused IOException

* Add method to write default directory to JSON

The default directory is used to open the explorer in the last directory
it was opened, when an image was added. The directory was not stored
permanently and afer every restart of the application the default
directory would be reset.

* Add method to get the default directory from JSON

* Remove unused imports and variables

* Fix code smells